Remove trailing slash from repository's home URL.
The second field of go-source meta tag is documented as:

> The home field is the URL of the repository's home page.

You're using the GitHub repository page as the home page URL. However, the
canonical URL for a GitHub repository does not have a trailing slash. E.g., if you
click on any repo from any link to it on GitHub, it won't have the trailing slash (but
explicitly navigating doesn't redirect).